Rename files and address comments
