+++ vendors/st/boards/stm32l475_discovery/aws_tests/config_files/iot_test_common_io_config.h
