+++ vendors/st/boards/stm32l475_discovery/ports/common_io/iot_test_common_io_internal.c
@@ -27,7 +27,7 @@
 #include "FreeRTOS.h"
 #include "task.h"
 
-#include "test_iot_internal.h"
+#include "iot_test_common_io_internal.h"
 
 /* UART */
 extern uint8_t uctestIotUartPort;
